Clean code with old functions and move IamClientRequest.java and WellKnownResponse.java files
